feat: support flexible output dtype for CuTe DSL MLA FP8 decode kernel
- Allow BFloat16 output for FP8 input (matching trtllm-gen backend default)
- FP16/BF16 input defaults to same dtype output; FP8 input defaults to BF16 output
- Add out_dtype parameter to cute_dsl_mla_decode for explicit override
- Add uses_shared_paged_kv_idx=False validation for cute-dsl backend
- Skip unsupported 3D page table tests for cute-dsl

+++ tests/attention/test_cute_dsl_mla_decode.py
@@ -448,7 +448,7 @@ def test_cute_dsl_mla_decode_fp8(batch_size, seq_len_k, page_size):
         output_scale=output_scale,
     )
 
-    assert out.dtype == torch.float8_e4m3fn
+    assert out.dtype == torch.bfloat16
     assert out.shape == (batch_size, q_len, num_heads, latent_dim)
 
     # Reference: compute in FP32 using FP8 values dequantized to FP32
